Changed 'enter' so that in normal rooms it doesn't ask 'to' anymore.  Now it only asks for a recipient in Mail.
Remove delete this; in button_goto (roomview.cpp), as it was causing
daphne to crash if goto was clicked more than once.  Unfortunately this
results in a potentially infinite number of room windows open at once,
but as this was the MAIN instability issue with the Linux version it had
to be dealt with.

Worked more on the windows port.  Executable is unstable and will be
worked on this week.  unfortunately wxWindows is not nearly as portable as
stated and I'll have to learn a whole new set of commands to deal with the
MSW port.  Getting there though.