Removed a few unneccesary lock calls to get a small speed up.
Changed the calc for load average. Now for the purposes of load averages
sleeping is the same as running since it means the thread can't do
something else in the mean time.

Started a new worker thread style for experimental purposes turn on by
un-commenting the #define NEW_WORKER at the top of threads.c