Now the autopurger will always attempt to purge users regardless of the "Default user purge time" but it will not purge users on time if it is disabled. This change added so that dead users are removed if need be.
Check this out guys. The changes are simple but lots of possible things
to test with catastrophic results if I got it wrong.

This should mean that users that have been deleted but not purged and
users that want to be deleted get deleted even if the user purge is
turned off.

Only deleteing users based on time is disabled if the user purge is
0.