Added code to report Aide messages if there appears to be something wrong with IGNet configuration. It now reports connection attempts from unknown nodes or bad passwords.
It also attempts to validate the nodename of the node it is attempting 
to connect to by checking the greeting message. In this case the 
connection will proceed with an Aide warning message if they differ. 
Situations such as this WILL result in duplication of messages but at 
least we now get a warning.