A little more with the API coding style.
Also added following functions

CtdlGetFloorByName   Gets floor by name. Return -1 if not found.
CtdlGetFloorByNameLock  As above but locks the floors.
CtdlGetAvailableFloor  Returns number of first unused floor.

These functions make use of the floor cache for performance.

The idea is if you want to create a floor you first check that the name
is available with CtdlGetFloorByNameLock
If the name is available you get -1 back and you then do CtdlGetAvailableFloor
Now you have a floor number to use so you can CtdlGetFloor and fill in the
structure information before doing CtdlPutFloorLock