CtdlRoomAccess() can now return a new value UA_REPLYALLOWED, which indicates that the user may post a message to this room *if* it is a reply to an existing message.
CtdlDoIHavePermissionToPostInThisRoom() now must be told whether
the user is replying to an existing message or posting a top-level
message.  If it is a reply, UA_REPLYALLOWED is checked.  If it is
a top-level message, UA_POSTALLOWED is checked.  If UA_POSTALLOWED
is not set, but UA_REPLYALLOWED is set, and the message is not a
reply, a new error message indicates that the user may only reply
to existing messages.  I may add a new error code for this later.