avoid calling GETR for the number of total/read messages
so far webcit used a call of GETR to retrieve the number of messages available in the current room for displaying the room banner.
However, GOTO already gave us this information, and GETR is only available for aides, so we got Important messages with that error when logged in as joe average.
This went under the radar for quiet a while, since the error message wasn't printed but silently ignored.